Major Overall update and fixes
Some work on the save mechanics allowing data transfer static files to be saved.

background for star map implemented. tried to do a repeating pattern but settled for static colour due to graphics

pop-up and confirmation window added when selecting a system to travel to. waiting for resource branch to be added for info

distance calculations fixed and ready for fuel implementation

fuel calculations added as well as capacity variables. ready for implementation of currency system and upgrades

camera script issues fixed with the ability to leave the parallax background and related issue of player leaving the camera behind

Zoom added to main camera in level to show speed and allow player to see in front of them properly.

Other minor fixes I've forgotten because I've been doing this all day

no need for separate branch as I have been working on the main and all scripts have been tested and completed in the same day